powerapp_yaml_package/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "powerapp-yaml"
dynamic = ["version"]
description = "Professional YAML handler for Power App components"
readme = "README.md"
license = {file = "LICENSE"}
authors = [{name = "PowerApp YAML Team", email = "team@powerapp-yaml.com"}]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
requires-python = ">=3.8"
dependencies = [
    "pyyaml>=6.0",
    "typing-extensions>=4.0.0; python_version<'3.10'",
]

[project.optional-dependencies]
dev = ["pytest>=6.0", "black", "flake8", "mypy", "pre-commit"]
docs = ["mkdocs", "mkdocs-material", "mkdocs-autoapi"]
test = ["pytest", "pytest-cov", "pytest-benchmark", "coverage"]

[project.urls]
Homepage = "https://github.com/powerapp/powerapp-yaml"
Documentation = "https://powerapp-yaml.readthedocs.io/"
Repository = "https://github.com/powerapp/powerapp-yaml"
"Bug Tracker" = "https://github.com/powerapp/powerapp-yaml/issues"

[project.scripts]
powerapp-yaml = "powerapp_yaml.cli:main"

[tool.setuptools_scm]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=powerapp_yaml --cov-report=html --cov-report=term"
